requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
reportlab==4.4.7
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2